+++ src/theme/palette.ts
@@ -1,0 +1,48 @@
+export default {
+  grey: {
+    50: '#f8f8f9',
+    100: '#eeeff1',
+    200: '#e3e4e7',
+    300: '#d7d9dd',
+    400: '#cfd0d6',
+    500: '#c6c8cf',
+    600: '#c0c2ca',
+    700: '#b9bbc3',
+    800: '#b1b4bd',
+    900: '#a4a7b2',
+    A100: '#ffffff',
+    A200: '#ffffff',
+    A400: '#f7f9ff',
+    A700: '#dee4ff',
+    contrastDefaultColor: 'dark',
+  },
+  mandarin: {
+    light: '#fff7cc',
+    main: '#ffd873',
+    dark: '#f3a110',
+  },
+  lavender: {
+    light: '#e8efff',
+    main: '#aebae4',
+    dark: '#636c8b',
+  },
+  cherry: {
+    light: '#ffe2df',
+    main: '#ff9f96',
+    dark: '#ff7467',
+  },
+  turquoise: {
+    light: '#e5fdf4',
+    main: '#8ae1d9',
+    dark: '#61beb5',
+  },
+  sand: {
+    main: '#fff8f0',
+  },
+  snow: {
+    main: '#ffffff',
+  },
+  ink: {
+    main: '#333333',
+  },
+};